changed from get to post on shopcart request and added cart component to index
