+++ src/components/Shopcart.js
@@ -4,47 +4,24 @@ import { Link, useOutletContext, useNavigate } from "react-router-dom";
 const Shopcart = () => {
     const { idState } = useOutletContext();
     const { productsState } = useOutletContext();
+    const { shopCartState } = useOutletContext();
     const { currentToken } = useOutletContext();
     const [userId, setUserId] = idState;
     const [products, setProducts] = productsState;
+    const [pendingOrders, setPendingOrders] = shopCartState;
     const navigate = useNavigate();
 
-    const [pendingOrders, setPendingOrders] = useState([]);
-
-    // api call for shopcart info based on standby
-    useEffect(() => {
-        async function loadPendingOrders() {
-            try {
-                const response = await fetch (`http://localhost:3030/api/shopcart/${userId}/status`, {
-                    headers: {
-                        "Content-Type": "application/json",
-                        "Authorization": `Bearer ${currentToken}`
-                    },
-                    body: JSON.stringify({
-                        cartStatus: "standby"
-                    })
-                })
-                const data = await response.json();
-                console.log("the order data: ", data);
-                setPendingOrders(data);
-            } catch (error) {
-                console.error(error)
-            }
-        }
-    
-        loadPendingOrders();
-
-    }, [])
-
-    // input for payment info
+    // import outlet context of open orders (standby)
+    // input for payment info  -- cardnumber, expiration date, cvv, address & name
     // drop down to change quantity
+    // button to remove item from cart
     // calculates total
     // button for checkout (verifies everything as accurate then takes you to a success page)
 
     return (
         <div>
             <p>shopping cart</p>
-            <Link to="/">Return to homepage</Link>
+            <Link to="products">Browse more items</Link>
         </div>
     )
 };
