+++ src/index.js
@@ -12,6 +12,7 @@ import Profile from "./components/Profile";
 import AdminPanel from "./components/AdminPanel";
 import Products from "./components/Products";
 import ProductDetails from "./components/ProductDetails";
+import Shopcart from "./components/Shopcart";
 
 const appElement = document.getElementById("app");
 const root = createRoot(appElement);
@@ -49,6 +50,10 @@ const router = createBrowserRouter([
             {
                 path: "/products/:id",
                 element: <ProductDetails />
+            },
+            {
+                path: "/shopcart",
+                element: <Shopcart />
             }
         ]
     }
